requirements.txt
aioredis==1.2.0
asgiref==3.2.1
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.8.1
Automat==0.7.0
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
Django==3.0.6
django-environ==0.4.5
django-filter==2.2.0
djangorestframework==3.11.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
Markdown==3.2.1
msgpack==0.6.1
psycopg2==2.8.3
pycparser==2.19
PyHamcrest==1.9.0
pylint==2.5.2
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
zope.interface==4.6.0
